refactor(code_gen): move tests into tests.rs
